adding toolbar spacing
