+++ src/pages/ClimateFeed.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Grid, makeStyles } from '@material-ui/core';
+import { Grid, makeStyles, Toolbar } from '@material-ui/core';
 import Loader from '../components/Loader';
 import CMCard from '../components/CMCard';
 import { useClimateFeed } from '../hooks/useClimateFeed';
@@ -23,43 +23,47 @@ const ClimateFeed: React.FC = () => {
     return <Loader />;
   }
   return (
-    <Grid
-      container
-      className={classes.root}
-      data-testid="ClimateFeed"
-      justify="space-around"
-    >
-      <Grid item sm={false} lg={4}>
-        {/* left gutter */}
-      </Grid>
-
+    <>
+      {/* Toolbar required to prevent content disap behind the app bar*/}
+      <Toolbar variant="dense" />
       <Grid
-        item
-        sm={12}
-        lg={4}
         container
-        direction="row"
-        justify="center"
-        alignItems="center"
+        className={classes.root}
+        data-testid="ClimateFeed"
+        justify="space-around"
       >
-        <Grid item sm={12} lg={12} container>
-          {climateFeed.map((effect, i) => (
-            <CMCard
-              key={`value-${i}`}
-              index={i}
-              title={effect.effectTitle}
-              shortDescription={effect.effectDescription}
-              numberedCards={false}
-              imageUrl={effect.imageUrl}
-            />
-          ))}
+        <Grid item sm={false} lg={4}>
+          {/* left gutter */}
+        </Grid>
+
+        <Grid
+          item
+          sm={12}
+          lg={4}
+          container
+          direction="row"
+          justify="center"
+          alignItems="center"
+        >
+          <Grid item sm={12} lg={12} container>
+            {climateFeed.map((effect, i) => (
+              <CMCard
+                key={`value-${i}`}
+                index={i}
+                title={effect.effectTitle}
+                shortDescription={effect.effectDescription}
+                numberedCards={false}
+                imageUrl={effect.imageUrl}
+              />
+            ))}
+          </Grid>
+        </Grid>
+
+        <Grid item sm={false} lg={4}>
+          {/* right gutter */}
         </Grid>
       </Grid>
-
-      <Grid item sm={false} lg={4}>
-        {/* right gutter */}
-      </Grid>
-    </Grid>
+    </>
   );
 };
